backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.0
qdrant-client>=1.7.0
groq>=0.4.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
trafilatura>=1.6.0
yt-dlp>=2023.12.0
beautifulsoup4>=4.12.0
